Validation and query for memory filtering. Refactored db method names

+++ api/db/schema.sql
@@ -3,14 +3,14 @@ CREATE TABLE IF NOT EXISTS Games (
     name TEXT NOT NULL,
     description TEXT,
     developer TEXT,
-    ram_min INTEGER,
+    ram_min FLOAT,
     cpu_min TEXT,
     gpu_min TEXT,
     OS_min TEXT,
-    storage_min INTEGER,
-    ram_rec INTEGER,
+    storage_min FLOAT,
+    ram_rec FLOAT,
     cpu_rec TEXT,
     gpu_rec TEXT,
     OS_rec TEXT,
-    storage_rec INTEGER
+    storage_rec FLOAT
 );
